log P25 call priority, mode, duplex
does not look at status bits in voice frames, only control channel messages

+++ trunk-recorder/systems/parser.h
@@ -39,7 +39,7 @@ struct TrunkMessage {
   bool emergency;
   bool duplex;
   bool mode;
-  long priority;
+  int priority;
   int tdma_slot;
   bool phase2_tdma;
   long source;
